requirements.txt
python-telegram-bot>=22.0
qbittorrent-api>=2024.1.55
httpx>=0.27
pyyaml>=6.0
